#  Use binary Heap to reduce number of key comparison in merging_iter