--- a/iteration/merging_iter.go
+++ b/iteration/merging_iter.go
@@ -3,23 +3,23 @@ package iteration
 import (
 	"bytes"
 	"github.com/spirit-labs/tektite/asl/encoding"
+	"math"
+
 	"github.com/spirit-labs/tektite/common"
 	log "github.com/spirit-labs/tektite/logger"
-	"math"
 )
 
 type MergingIterator struct {
 	highestVersion             uint64
 	iters                      []Iterator
-	dropIterCurrent            []bool
 	preserveTombstones         bool
-	current                    common.KV
-	currIndex                  int
 	minNonCompactableVersion   uint64
 	currentTombstone           []byte
 	isPrefixTombstone          bool
 	firstSameKey               []byte
 	firstSameKeyNonCompactable bool
+	heap                       []int
+	heapTail                   int
 }
 
 func NewMergingIterator(iters []Iterator, preserveTombstones bool, highestVersion uint64) (*MergingIterator, error) {
@@ -27,8 +27,12 @@ func NewMergingIterator(iters []Iterator, preserveTombstones bool, highestVersio
 		highestVersion:           highestVersion,
 		minNonCompactableVersion: math.MaxUint64,
 		iters:                    iters,
-		dropIterCurrent:          make([]bool, len(iters)),
 		preserveTombstones:       preserveTombstones,
+		heap:                     make([]int, len(iters)+1),
+		heapTail:                 0,
+	}
+	for i := 0; i < len(mi.heap); i++ {
+		mi.heap[i] = i - 1
 	}
 	return mi, nil
 }
@@ -38,149 +42,218 @@ func NewCompactionMergingIterator(iters []Iterator, preserveTombstones bool, min
 		highestVersion:           math.MaxUint64,
 		minNonCompactableVersion: minNonCompactableVersion,
 		iters:                    iters,
-		dropIterCurrent:          make([]bool, len(iters)),
 		preserveTombstones:       preserveTombstones,
+		heap:                     make([]int, len(iters)+1),
+		heapTail:                 0,
+	}
+	for i := 0; i < len(mi.heap); i++ {
+		mi.heap[i] = i - 1
 	}
 	return mi, nil
 }
 
 func (m *MergingIterator) Next() (bool, common.KV, error) {
+	var err error
+	var valid bool
 
 	var lastKeyNoVersion []byte
-	if len(m.current.Key) > 0 {
+	var lastKeyVersion uint64
+	var lastKV common.KV
+	if m.heap[0] >= 0 {
+		lastKV = m.iters[m.heap[0]].Current()
 		// If a key is not compactable because version >= minNonCompactableVersion then all versions of that key
 		// are also non compactable, even if the other versions are < minNonCompactableVersion
 		// so, we need to keep track of when the last key changes then decide whether that key is compactable
 		// then we won't skip past any entries that are non compactable
-		lastKeyNoVersion = m.current.Key[:len(m.current.Key)-8]
+		lastKeyNoVersion = lastKV.Key[:len(lastKV.Key)-8]
+		lastKeyVersion = encoding.DecodeKeyVersion(lastKV.Key)
 		if m.firstSameKey == nil || !bytes.Equal(m.firstSameKey, lastKeyNoVersion) {
 			m.firstSameKey = lastKeyNoVersion
-			m.firstSameKeyNonCompactable = encoding.DecodeKeyVersion(m.current.Key) >= m.minNonCompactableVersion
+			m.firstSameKeyNonCompactable = lastKeyVersion >= m.minNonCompactableVersion
 		}
 	}
 
-	repeat := true
-	for repeat {
-		// Now find the smallest key from all iterators, choosing the highest version when keys draw
-		var keyNoVersion, chosenKeyNoVersion []byte
-		var smallestIndex int
-		var chosen common.KV
-		var version, chosenVersion uint64
-		var err error
-		var valid bool
-	outer:
-		for i := range m.iters {
-			c := m.iters[i].Current()
-			if len(c.Key) == 0 || m.dropIterCurrent[i] {
-				for {
-					// Find the next valid iters KV that is compactable and has a version that is not too high
-					valid, c, err = m.iters[i].Next()
-					if err != nil {
-						return false, common.KV{}, err
-					}
-					if !valid {
-						continue outer
-					}
+	advanceHeapIter := -1
+	inactiveIterToCheck := -1
 
-					version = encoding.DecodeKeyVersion(c.Key)
-					// Skip over same key (if it's compactable)
-					// in same iterator we can have multiple versions of the same key
-					keyNoVersion = c.Key[:len(c.Key)-8]
-					if !m.firstSameKeyNonCompactable && bytes.Equal(lastKeyNoVersion, keyNoVersion) {
-						if version >= m.minNonCompactableVersion {
-							panic("compacting non compactable version") //sanity check
-						}
-						if log.DebugEnabled {
-							lastVersion := encoding.DecodeKeyVersion(m.current.Key)
-							log.Debugf("%p mi: dropping key in next as same key: key %v (%s) value %v (%s) version:%d last key: %v (%s) last value %v (%s) last version %d minnoncompactableversion:%d",
-								m, c.Key, string(c.Key), c.Value, string(c.Value), version, m.current.Key, string(m.current.Key), m.current.Value, string(m.current.Value), lastVersion, m.minNonCompactableVersion)
-						}
-						continue
-					}
-					// Skip past keys with too high a version
-					// prefix tombstones always have version math.MaxUint64 and are never screened out
-					if version > m.highestVersion && version != math.MaxUint64 {
-						if log.DebugEnabled {
-							log.Debugf("%p merging iter skipping past key %v (%s) as version %d too high - max version %d",
-								m, c.Key, string(c.Key), version, m.highestVersion)
-						}
-						continue
-					}
-					m.dropIterCurrent[i] = false
+	if m.heap[0] != -1 && m.heapTail >= 1 {
+		// Replace/Remove the last returned value if the heap is not empty
+		advanceHeapIter = 1
+	}
+
+	var chosenVersion uint64
+	var chosenKeyNoVersion []byte
+	var kv common.KV
+	for {
+		if advanceHeapIter >= 1 {
+			for {
+				valid, kv, err = m.iters[m.heap[advanceHeapIter]].Next()
+				if err != nil {
+					return false, kv, err
+				}
+				if !valid {
 					break
 				}
-			} else {
-				version = encoding.DecodeKeyVersion(c.Key)
-				keyNoVersion = c.Key[:len(c.Key)-8]
+				version := encoding.DecodeKeyVersion(kv.Key)
+				// Skip over same key (if it's compactable)
+				// in same iterator we can have multiple versions of the same key
+				keyNoVersion := kv.Key[:len(kv.Key)-8]
+				if !m.firstSameKeyNonCompactable && bytes.Equal(lastKeyNoVersion, keyNoVersion) {
+					if version >= m.minNonCompactableVersion {
+						panic("compacting non compactable version") //sanity check
+					}
+					if log.DebugEnabled {
+						lastVersion := encoding.DecodeKeyVersion(lastKV.Key)
+						log.Debugf("%p mi: dropping key in next as same key: key %v (%s) value %v (%s) version:%d last key: %v (%s) last value %v (%s) last version %d minnoncompactableversion:%d",
+							m, kv.Key, string(kv.Key), kv.Value, string(kv.Value), version, lastKV.Key, string(lastKV.Key), lastKV.Value, string(lastKV.Value), lastVersion, m.minNonCompactableVersion)
+					}
+					continue
+				}
+				// Skip past keys with too high a version
+				// prefix tombstones always have version math.MaxUint64 and are never screened out
+				if version > m.highestVersion && version != math.MaxUint64 {
+					if log.DebugEnabled {
+						log.Debugf("%p merging iter skipping past key %v (%s) as version %d too high - max version %d",
+							m, kv.Key, string(kv.Key), version, m.highestVersion)
+					}
+					continue
+				}
+				break
 			}
 
-			if chosenKeyNoVersion == nil {
-				chosenKeyNoVersion = keyNoVersion
-				smallestIndex = i
-				chosen = c
-				chosenVersion = version
-			} else {
-				diff := bytes.Compare(keyNoVersion, chosenKeyNoVersion)
-				if diff < 0 {
-					chosenKeyNoVersion = keyNoVersion
-					smallestIndex = i
-					chosen = c
-					chosenVersion = version
-				} else if diff == 0 {
-					// Keys are same
-
-					// choose the highest version, and drop the other one as long as the highest version < minNonCompactable
-					if version > chosenVersion {
-						// the current version is higher so drop the previous highest if the current version is compactable
-						// note we can only drop the previous highest if *this* version is compactable as dropping it
-						// will leave this version, and if its non compactable it means that snapshot rollback could
-						// remove it, which would leave nothing.
-						if version < m.minNonCompactableVersion {
-							if log.DebugEnabled {
-								log.Debugf("%p mi: dropping as key version %d less than minnoncompactable (%d) %d chosenKeyVersion %d: key %v (%s) value %v (%s)",
-									m, version, 1, m.minNonCompactableVersion, chosenVersion, c.Key, string(c.Key), c.Value, string(c.Value))
-							}
-							m.dropIterCurrent[smallestIndex] = true
+			if advanceHeapIter > m.heapTail {
+				// Invalid iterator advanced. If it became valid, then move it to the end of valid region
+				// and sifting it up
+				if valid {
+					m.heapTail++ // make room to the new valid kv at the end of the heap.
+					if advanceHeapIter != m.heapTail {
+						// if iter not in the tail, move it to the tail
+						m.heap[advanceHeapIter], m.heap[m.heapTail] = m.heap[m.heapTail], m.heap[advanceHeapIter]
+					}
+					// Heap Sift Up operation
+					child := m.heapTail
+					for {
+						parent := child / 2
+						if parent < 1 {
+							break
 						}
-						chosenKeyNoVersion = keyNoVersion
-						smallestIndex = i
-						chosen = c
-						chosenVersion = version
-					} else if version < chosenVersion {
-						// the previous highest version is higher than this version, so we can remove this version
-						// as long as previous highest is compactable
-						if chosenVersion < m.minNonCompactableVersion {
-							// drop this entry if the version is compactable
-							if log.DebugEnabled {
-								log.Debugf("%p mi: dropping as key version %d less than minnoncompactable (%d) %d chosenKeyVersion %d: key %v (%s) value %v (%s)",
-									m, version, 2, m.minNonCompactableVersion, chosenVersion, chosen.Key, string(chosen.Key), chosen.Value, string(chosen.Value))
-							}
-							m.dropIterCurrent[i] = true
+						parentKV := m.iters[m.heap[parent]].Current()
+						childKV := m.iters[m.heap[child]].Current()
+						diff := bytes.Compare(childKV.Key[:len(childKV.Key)-8], parentKV.Key[:len(parentKV.Key)-8])
+						if diff < 0 || (diff == 0 && encoding.DecodeKeyVersion(childKV.Key) > encoding.DecodeKeyVersion(parentKV.Key)) {
+							// child is smaller, swap parent-child and keep sifting up.
+							m.heap[parent], m.heap[child] = m.heap[child], m.heap[parent]
+							child = parent
+						} else {
+							break
 						}
-					} else {
-						// same key, same version, drop this one, and keep the one we already found,
-						if log.DebugEnabled {
-							log.Debugf("%p mi: dropping key as same key and version: key %v (%s) value %v (%s) ver %d",
-								m, c.Key, string(c.Key), c.Value, string(c.Value), version)
-						}
-						m.dropIterCurrent[i] = true
 					}
 				}
+			} else if advanceHeapIter == 1 {
+				siftDownHead := false
+				if valid {
+					siftDownHead = true
+				} else {
+					// Valid iterator advanced. If it became invalid, then move it to the invalid region
+					if m.heapTail >= 2 {
+						m.heap[1], m.heap[m.heapTail] = m.heap[m.heapTail], m.heap[1]
+					}
+					m.heapTail--
+					if m.heapTail >= 2 {
+						siftDownHead = true
+					}
+				}
+				if siftDownHead {
+					// Heap Sift Down operation
+					parent := 1
+					smallest := parent
+					for {
+						smallestKV := m.iters[m.heap[smallest]].Current()
+						smallestKeyWithoutVersion := smallestKV.Key[:len(smallestKV.Key)-8]
+						smallestKeyVersion := encoding.DecodeKeyVersion(smallestKV.Key)
+
+						for _, child := range [2]int{2 * parent, 2*parent + 1} {
+							if child > m.heapTail {
+								break
+							}
+							childKV := m.iters[m.heap[child]].Current()
+							childKeyWithoutVersion := childKV.Key[:len(childKV.Key)-8]
+							childKeyVersion := encoding.DecodeKeyVersion(childKV.Key)
+							diff := bytes.Compare(childKeyWithoutVersion, smallestKeyWithoutVersion)
+
+							// If keys are same, the smallest should be the one with the highest version
+							// if version are the same, sort it based on iter index
+							if diff < 0 || (diff == 0 && (childKeyVersion > smallestKeyVersion || (childKeyVersion == smallestKeyVersion &&
+								m.heap[child] < m.heap[smallest]))) {
+								smallest = child
+								smallestKeyWithoutVersion = childKeyWithoutVersion
+								smallestKeyVersion = childKeyVersion
+							}
+						}
+
+						// parent is smallest key amongst its children - heap variant is satisfied
+						if smallest == parent {
+							break
+						}
+						// swap
+						m.heap[smallest], m.heap[parent] = m.heap[parent], m.heap[smallest]
+						parent = smallest // keep sinking down prev smaller
+					}
+				}
+			} else {
+				panic("Only allowed to advance iterator in the HEAD or in the invalid region")
 			}
 		}
 
-		if chosenKeyNoVersion == nil {
+		if inactiveIterToCheck == -1 { // if it hasn't check inactive iters, it starts from the first one
+			inactiveIterToCheck = m.heapTail + 1
+			if advanceHeapIter == 1 && !valid {
+				// if a prev valid iter just became invalid, we don't need to check it. so we skip the first invalid
+				inactiveIterToCheck++
+			}
+		} else {
+			inactiveIterToCheck++
+		}
+		if inactiveIterToCheck < len(m.heap) {
+			// keep checking inactive iterator
+			advanceHeapIter = inactiveIterToCheck
+			continue
+		}
+
+		if m.heapTail < 1 {
 			// Nothing valid
-			m.current = common.KV{}
+			m.heap[0] = -1
 			return false, common.KV{}, nil
 		}
 
+		choosenKV := m.iters[m.heap[1]].Current()
+		chosenKeyNoVersion = choosenKV.Key[:len(choosenKV.Key)-8]
+		chosenVersion = encoding.DecodeKeyVersion(choosenKV.Key)
+
+		if bytes.Equal(chosenKeyNoVersion, m.firstSameKey) {
+			// same key, same version as last returned, drop this one
+			if chosenVersion == lastKeyVersion {
+				advanceHeapIter = 1
+				continue
+			}
+			// the previous highest version is higher than this version, so we can remove this version
+			// as long as previous highest is compactable
+			if !m.firstSameKeyNonCompactable {
+				if log.DebugEnabled {
+					log.Debugf("%p mi: dropping as key version %d less than minnoncompactable (%d) %d chosenKeyVersion %d: key %v (%s) value %v (%s)",
+						m, lastKeyVersion, 1, m.minNonCompactableVersion, chosenVersion, lastKV.Key, lastKV.Key, lastKV.Value, string(lastKV.Value))
+				}
+				advanceHeapIter = 1
+				continue
+			}
+		}
+
 		if m.currentTombstone != nil {
-			if bytes.Equal(m.currentTombstone, chosen.Key[:len(m.currentTombstone)]) {
+			if bytes.Equal(m.currentTombstone, choosenKV.Key[:len(m.currentTombstone)]) {
 				if m.isPrefixTombstone || chosenVersion < m.minNonCompactableVersion {
 					// The key matches current prefix tombstone
-					// skip past it if it is compactable - for prefixes we alwqays delete even if non compactable
-					m.dropIterCurrent[smallestIndex] = true
+					// skip past it if it is compactable - for prefixes we alwqys delete even if non compactable
+					advanceHeapIter = 1
 					continue
 				}
 			} else {
@@ -190,7 +263,7 @@ func (m *MergingIterator) Next() (bool, common.KV, error) {
 			}
 		}
 
-		isTombstone := len(chosen.Value) == 0
+		isTombstone := len(choosenKV.Value) == 0
 		if isTombstone {
 			// We have a tombstone, keep track of it. Prefix tombstones (used for deletes of partitions)
 			// are identified by having a version of math.MaxUint64
@@ -202,21 +275,20 @@ func (m *MergingIterator) Next() (bool, common.KV, error) {
 		if !m.preserveTombstones && (isTombstone || chosenVersion == math.MaxUint64) {
 			// We have a tombstone or a prefix tombstone end marker - skip past it
 			// End marker also is identified as having a version of math.MaxUint64
-			m.dropIterCurrent[smallestIndex] = true
-		} else {
-			// output the entry
-			m.current.Key = chosen.Key
-			m.current.Value = chosen.Value
-			m.currIndex = smallestIndex
-			repeat = false
+			advanceHeapIter = 1
+			continue
 		}
+
+		m.heap[0] = m.heap[1]
+		return true, choosenKV, nil
 	}
-	m.dropIterCurrent[m.currIndex] = true
-	return true, m.current, nil
 }
 
 func (m *MergingIterator) Current() common.KV {
-	return m.current
+	if m.heap[0] < 0 {
+		return common.KV{}
+	}
+	return m.iters[m.heap[0]].Current()
 }
 
 func (m *MergingIterator) Close() {