@purplefox 2024-08-11
Looks good.

A couple of suggestions / thoughts:

* Could you run the benchmark for an extended time (e.g. 5 mins, with -benchtime=300s)? - that makes it more likely that GC occurs. With short running benchmarks GC often does not occur, so if the code is creating a lot of garbage it won't be reflected in the performance figures. Extending the time makes the benchmark behave more like it would under sustained load
* How likely is it that we would have a merging iterator with more than 10 iterators? Typically when we create queries or do compaction, which uses merging iterators, we would have one iterator per level in the LSM, so we'd have 1, 2, 3, 4 or maybe 5 iterators for a very large database? It would be good to see benchmark results for 2...5 iterators as these are going to be most common numbers.