@watson28 2024-08-14
> Could you run the benchmark for an extended time (e.g. 5 mins, with -benchtime=300s)? - that makes it more likely that GC occurs. With short running benchmarks GC often does not occur, so if the code is creating a lot of garbage it won't be reflected in the performance figures. Extending the time makes the benchmark behave more like it would under sustained load

@purplefox I ran the the benchmark with 5 mins time and used a single benchmark test with 10 iterators (as in main), and this it's what I got:

old:
``` 
go test -benchmem -bench ^BenchmarkMergingIterator$ github.com/spirit-labs/tektite/iteration -benchtime=5m
goos: darwin
goarch: arm64
pkg: github.com/spirit-labs/tektite/iteration
BenchmarkMergingIterator-10    	 1956621	    184049 ns/op	     208 B/op	       2 allocs/op
PASS
ok  	github.com/spirit-labs/tektite/iteration	546.216s
```

new:
```
go test -benchmem -bench ^BenchmarkMergingIterator$ github.com/spirit-labs/tektite/iteration -benchtime=5m
goos: darwin
goarch: arm64
pkg: github.com/spirit-labs/tektite/iteration
BenchmarkMergingIterator-10    	 2389642	    151641 ns/op	     240 B/op	       2 allocs/op
PASS
ok  	github.com/spirit-labs/tektite/iteration	514.828s
```

is there an specific aspect of the benchmark to check for gc load or the execution time is enough?

>How likely is it that we would have a merging iterator with more than 10 iterators? Typically when we create queries or do compaction, which uses merging iterators, we would have one iterator per level in the LSM, so we'd have 1, 2, 3, 4 or maybe 5 iterators for a very large database? It would be good to see benchmark results for 2...5 iterators as these are going to be most common numbers.

you are right. I mix-up with the scenario where queries need to scan a large range, but we only need to merge data between levels. In a single level, SSTs don't overlap so merging doesn't happen between SSTs in a level. And I assume overlap also doesn't occur in level 0 (I remember seeing a pr regarding this). 
I would update the benchmark cases to have iterators from 2 to 5.

May I ask what criteria we use to merge and move tables to the next level? Is it when a level reaches X number of SSTs?