@purplefox 2024-08-15
> And I assume overlap also doesn't occur in level 0

Level 0 is split into multiple groups - one for each processor, and there is no overlap inside a particular group. However when executing a query we always execute them for a particular partition which will map to a single L0 group. Also when merging a compaction job only ever has one group from L0 in it, so we can parallelise compaction. So yes, we only ever have one iterator from L0 when merging or querying.

> May I ask what criteria we use to merge and move tables to the next level? Is it when a level reaches X number of SSTs?

Yes, we choose the level which exceeds the level ideal size by the greatest amount.