@watson28 2024-08-19
@purplefox I run the benchmark with the low number of iterators: 2,3,4,5 and this is what I got

main:
```
goos: darwinx
goarch: arm64
pkg: github.com/spirit-labs/tektite/iteration
BenchmarkMergingIterator/MerginIterator-2-10         	   22350	     52554 ns/op	     194 B/op	       2 allocs/op
BenchmarkMergingIterator/MerginIterator-3-10         	   18362	     65015 ns/op	     195 B/op	       2 allocs/op
BenchmarkMergingIterator/MerginIterator-4-10         	   15324	     78920 ns/op	     196 B/op	       2 allocs/op
BenchmarkMergingIterator/MerginIterator-5-10         	   12031	     99560 ns/op	     197 B/op	       2 allocs/op
PASS
ok  	github.com/spirit-labs/tektite/iteration	7.986s
```

this branch:
```
goos: darwin
goarch: arm64
pkg: github.com/spirit-labs/tektite/iteration
BenchmarkMergingIterator/MerginIterator-2-10         	   15562	     76435 ns/op	     168 B/op	       2 allocs/op
BenchmarkMergingIterator/MerginIterator-3-10         	   13046	     91853 ns/op	     176 B/op	       2 allocs/op
BenchmarkMergingIterator/MerginIterator-4-10         	   10000	    103256 ns/op	     192 B/op	       2 allocs/op
BenchmarkMergingIterator/MerginIterator-5-10         	    9970	    120007 ns/op	     192 B/op	       2 allocs/op
PASS
ok  	github.com/spirit-labs/tektite/iteration	6.567s
```

the current implementation is faster than the one with a heap. I guess having a heap only paid off with large number of iterators to merge but as you said this will never be the case for the merging_iter.
So, I think I will close this PR and open a new one with only the changes I made in the benchmark test: run tests with 2,3,4 y 5 iterators.