@purplefox 2024-08-20
Makes sense. Thank you for looking at this anyway.